Decreases the bridge flow ratio
This results in thinner extrudes for bridging. There is still some extrude lines that stick out causing potential nozzle collisions over the next layer, but as the bridge extrude line is much more thinner these bumps are a bit more flexible and a bit more easy for the nozzle to go over. Not to mention they seem less pronounced.